+++ cms-templates/src/main/java/com/condation/cms/templates/components/TemplateComponents.java
@@ -25,11 +25,9 @@ package com.condation.cms.templates.components;
 
 import com.condation.cms.api.model.Parameter;
 import com.condation.cms.api.request.RequestContext;
-import java.util.Collections;
 import java.util.Map;
 import java.util.Set;
 import java.util.function.Function;
-import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 /**